requirements.txt
mcp
python-dotenv
google-genai
Pillow
# pywinauto
asyncio 